refactor: separate archive transport and persistence modules
