+++ snapshot/snapshot.go
@@ -1,24 +1,9 @@
 package snapshot
 
 import (
-	"bufio"
-	"compress/gzip"
 	"context"
-	"crypto/sha256"
 	"database/sql"
-	"encoding/hex"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"hash"
-	"io"
-	"os"
-	"path/filepath"
-	"sort"
-	"strings"
 	"time"
-
-	"github.com/openclaw/crawlkit/store"
 )
 
 const ManifestName = "manifest.json"
@@ -70,69 +55,6 @@ type ImportProgress struct {
 	TotalRows int
 }
 
-type Sidecar struct {
-	Name   string `json:"name"`
-	Path   string `json:"path"`
-	Kind   string `json:"kind,omitempty"`
-	Size   int64  `json:"size,omitempty"`
-	SHA256 string `json:"sha256,omitempty"`
-}
-
-type Manifest struct {
-	Version     int               `json:"version"`
-	GeneratedAt time.Time         `json:"generated_at"`
-	Tables      []TableManifest   `json:"tables"`
-	Sidecars    []Sidecar         `json:"sidecars,omitempty"`
-	Files       map[string]string `json:"files,omitempty"`
-}
-
-type TableManifest struct {
-	Name          string         `json:"name"`
-	File          string         `json:"file,omitempty"`
-	Files         []string       `json:"files"`
-	FileManifests []FileManifest `json:"file_manifests,omitempty"`
-	Columns       []string       `json:"columns"`
-	Rows          int            `json:"rows"`
-}
-
-type FileManifest struct {
-	Path   string `json:"path"`
-	Rows   int    `json:"rows"`
-	Size   int64  `json:"size,omitempty"`
-	SHA256 string `json:"sha256,omitempty"`
-}
-
-var ErrNoManifest = errors.New("pack manifest not found")
-
-type TableImportMode string
-
-const (
-	TableImportSkip    TableImportMode = "skip"
-	TableImportReplace TableImportMode = "replace"
-	TableImportFiles   TableImportMode = "files"
-)
-
-type ImportPlan struct {
-	Full   bool
-	Reason string
-	Tables []TableImportPlan
-}
-
-type ImportImpact string
-
-const (
-	ImportImpactNone    ImportImpact = "none"
-	ImportImpactMerge   ImportImpact = "merge"
-	ImportImpactReplace ImportImpact = "replace"
-)
-
-type TableImportPlan struct {
-	Table  TableManifest
-	Mode   TableImportMode
-	Files  []FileManifest
-	Reason string
-}
-
 type IncrementalImportOptions struct {
 	DB           *sql.DB
 	RootDir      string
@@ -149,870 +71,4 @@ type IncrementalImportOptions struct {
 
 func Export(ctx context.Context, opts ExportOptions) (Manifest, error) {
 	return exportSnapshot(ctx, opts)
-}
-
-func Import(ctx context.Context, opts ImportOptions) (Manifest, error) {
-	if opts.DB == nil {
-		return Manifest{}, errors.New("db is required")
-	}
-	manifest, err := ReadManifest(opts.RootDir)
-	if err != nil {
-		return Manifest{}, err
-	}
-	deleteTables := opts.DeleteTables
-	if len(deleteTables) == 0 {
-		for _, table := range manifest.Tables {
-			deleteTables = append(deleteTables, table.Name)
-		}
-	}
-	tx, err := opts.DB.BeginTx(ctx, nil)
-	if err != nil {
-		return Manifest{}, fmt.Errorf("begin import tx: %w", err)
-	}
-	committed := false
-	defer func() {
-		if !committed {
-			_ = tx.Rollback()
-		}
-	}()
-	if opts.BeforeImport != nil {
-		if err := opts.BeforeImport(ctx, tx); err != nil {
-			return Manifest{}, err
-		}
-	}
-	for i := len(deleteTables) - 1; i >= 0; i-- {
-		table := strings.TrimSpace(deleteTables[i])
-		if table == "" {
-			continue
-		}
-		if opts.DeleteTable != nil {
-			if err := opts.DeleteTable(ctx, tx, table); err != nil {
-				return Manifest{}, err
-			}
-			continue
-		}
-		if _, err := tx.ExecContext(ctx, "delete from "+store.QuoteIdent(table)); err != nil {
-			return Manifest{}, fmt.Errorf("clear table %s: %w", table, err)
-		}
-	}
-	for _, table := range manifest.Tables {
-		rows, err := importTable(ctx, tx, opts.RootDir, table, opts.Filter, opts.ImportRow, opts.Progress)
-		if err != nil {
-			return Manifest{}, err
-		}
-		reportImportProgress(opts.Progress, ImportProgress{Phase: "table_done", Table: table.Name, Rows: rows, TotalRows: table.Rows})
-	}
-	if opts.AfterImport != nil {
-		if err := opts.AfterImport(ctx, tx); err != nil {
-			return Manifest{}, err
-		}
-	}
-	if err := tx.Commit(); err != nil {
-		return Manifest{}, fmt.Errorf("commit import tx: %w", err)
-	}
-	committed = true
-	return manifest, nil
-}
-
-func PlanIncrementalImport(previous, current Manifest) ImportPlan {
-	return planIncrementalImport(previous, current, false)
-}
-
-// PlanMergeImport prepares a monotonic cache refresh. Changed snapshot files
-// are upserted without deleting rows that disappeared from the snapshot. Use
-// PlanIncrementalImport when the destination must exactly mirror the source.
-func PlanMergeImport(previous, current Manifest) ImportPlan {
-	return planIncrementalImport(previous, current, true)
-}
-
-func planIncrementalImport(previous, current Manifest, merge bool) ImportPlan {
-	if current.Version != previous.Version {
-		return ImportPlan{Full: true, Reason: "manifest version changed"}
-	}
-	previousTables := make(map[string]TableManifest, len(previous.Tables))
-	for _, table := range previous.Tables {
-		previousTables[table.Name] = table
-	}
-	currentTables := make(map[string]TableManifest, len(current.Tables))
-	for _, table := range current.Tables {
-		currentTables[table.Name] = table
-	}
-	for name := range previousTables {
-		if _, ok := currentTables[name]; !ok {
-			return ImportPlan{Full: true, Reason: "table removed: " + name}
-		}
-	}
-	plan := ImportPlan{}
-	for _, table := range current.Tables {
-		previousTable, ok := previousTables[table.Name]
-		if !ok {
-			mode := TableImportReplace
-			reason := "new table"
-			if merge {
-				mode = TableImportFiles
-				reason = "merge new table"
-			}
-			plan.Tables = append(plan.Tables, TableImportPlan{
-				Table:  table,
-				Mode:   mode,
-				Files:  tableFileManifests(table),
-				Reason: reason,
-			})
-			continue
-		}
-		tablePlan := planTableIncrement(previousTable, table, merge)
-		plan.Tables = append(plan.Tables, tablePlan)
-	}
-	return plan
-}
-
-func (p ImportPlan) Changed() bool {
-	if p.Full {
-		return true
-	}
-	for _, table := range p.Tables {
-		if table.Mode != TableImportSkip {
-			return true
-		}
-	}
-	return false
-}
-
-func (p ImportPlan) Impact() ImportImpact {
-	if p.Full {
-		return ImportImpactReplace
-	}
-	impact := ImportImpactNone
-	for _, table := range p.Tables {
-		switch table.Mode {
-		case TableImportReplace:
-			return ImportImpactReplace
-		case TableImportFiles:
-			impact = ImportImpactMerge
-		}
-	}
-	return impact
-}
-
-func ImportIncremental(ctx context.Context, opts IncrementalImportOptions) (Manifest, ImportPlan, error) {
-	if opts.DB == nil {
-		return Manifest{}, ImportPlan{}, errors.New("db is required")
-	}
-	current := opts.Current
-	var err error
-	if len(current.Tables) == 0 {
-		current, err = ReadManifest(opts.RootDir)
-		if err != nil {
-			return Manifest{}, ImportPlan{}, err
-		}
-	}
-	currentTables := make(map[string]TableManifest, len(current.Tables))
-	for _, table := range current.Tables {
-		if _, exists := currentTables[table.Name]; exists {
-			return Manifest{}, ImportPlan{}, fmt.Errorf("duplicate table %q in the current manifest", table.Name)
-		}
-		if _, err := importFileManifests(table); err != nil {
-			return Manifest{}, ImportPlan{}, err
-		}
-		currentTables[table.Name] = table
-	}
-	plan := opts.Plan
-	if len(plan.Tables) == 0 && !plan.Full && plan.Reason == "" {
-		plan = PlanIncrementalImport(opts.Previous, current)
-	}
-	if plan.Full {
-		return Manifest{}, plan, errors.New("incremental import requires a non-full plan: " + plan.Reason)
-	}
-	if !plan.Changed() {
-		return current, plan, nil
-	}
-	activeTables := make(map[string]bool, len(plan.Tables))
-	for _, tablePlan := range plan.Tables {
-		if tablePlan.Mode == TableImportSkip {
-			continue
-		}
-		if activeTables[tablePlan.Table.Name] {
-			return Manifest{}, plan, fmt.Errorf("duplicate active plan for table %q", tablePlan.Table.Name)
-		}
-		activeTables[tablePlan.Table.Name] = true
-	}
-	work, err := prepareIncrementalWork(plan, currentTables)
-	if err != nil {
-		return Manifest{}, plan, err
-	}
-	tx, err := opts.DB.BeginTx(ctx, nil)
-	if err != nil {
-		return Manifest{}, plan, fmt.Errorf("begin incremental import tx: %w", err)
-	}
-	committed := false
-	defer func() {
-		if !committed {
-			_ = tx.Rollback()
-		}
-	}()
-	if err := checkIncrementalDependencies(ctx, tx, work, opts); err != nil {
-		return Manifest{}, plan, err
-	}
-	if opts.BeforeImport != nil {
-		if err := opts.BeforeImport(ctx, tx); err != nil {
-			return Manifest{}, plan, err
-		}
-	}
-	for _, item := range work {
-		table := item.table
-		if item.mode == TableImportReplace {
-			if err := deleteImportTable(ctx, tx, table.Name, opts.DeleteTable); err != nil {
-				return Manifest{}, plan, err
-			}
-		}
-		rows, err := importTable(ctx, tx, opts.RootDir, table, opts.Filter, opts.ImportRow, opts.Progress)
-		if err != nil {
-			return Manifest{}, plan, err
-		}
-		reportImportProgress(opts.Progress, ImportProgress{Phase: "table_done", Table: table.Name, Rows: rows, TotalRows: table.Rows})
-	}
-	if opts.AfterImport != nil {
-		if err := opts.AfterImport(ctx, tx); err != nil {
-			return Manifest{}, plan, err
-		}
-	}
-	if err := tx.Commit(); err != nil {
-		return Manifest{}, plan, fmt.Errorf("commit incremental import tx: %w", err)
-	}
-	committed = true
-	return current, plan, nil
-}
-
-func ReadManifest(rootDir string) (Manifest, error) {
-	data, err := os.ReadFile(filepath.Join(rootDir, ManifestName))
-	if errors.Is(err, os.ErrNotExist) {
-		return Manifest{}, ErrNoManifest
-	}
-	if err != nil {
-		return Manifest{}, err
-	}
-	var manifest Manifest
-	if err := json.Unmarshal(data, &manifest); err != nil {
-		return Manifest{}, fmt.Errorf("parse manifest: %w", err)
-	}
-	return manifest, nil
-}
-
-func WriteManifest(rootDir string, manifest Manifest) error {
-	data, err := json.MarshalIndent(manifest, "", "  ")
-	if err != nil {
-		return fmt.Errorf("marshal manifest: %w", err)
-	}
-	data = append(data, '\n')
-	if err := os.MkdirAll(rootDir, 0o755); err != nil {
-		return fmt.Errorf("create root dir: %w", err)
-	}
-	if err := os.WriteFile(filepath.Join(rootDir, ManifestName), data, 0o600); err != nil {
-		return fmt.Errorf("write manifest: %w", err)
-	}
-	return nil
-}
-
-func exportTable(ctx context.Context, tx *sql.Tx, root *os.Root, generation, table string, maxShardBytes int64, filter RowFilter, filterTx RowFilterTx, created func(string)) (TableManifest, error) {
-	relDir, err := tableShardDir(table)
-	if err != nil {
-		return TableManifest{}, err
-	}
-	relDir = generation + "/" + table
-	rows, err := tx.QueryContext(ctx, "select * from "+store.QuoteIdent(table))
-	if err != nil {
-		return TableManifest{}, fmt.Errorf("query table %s: %w", table, err)
-	}
-	defer rows.Close()
-	cols, err := rows.Columns()
-	if err != nil {
-		return TableManifest{}, err
-	}
-	writer := &shardWriter{
-		root:          root,
-		relDir:        relDir,
-		maxShardBytes: maxShardBytes,
-		created:       created,
-	}
-	if err := root.MkdirAll(filepath.FromSlash(relDir), 0o755); err != nil {
-		return TableManifest{}, fmt.Errorf("create table dir %s: %w", table, err)
-	}
-	defer writer.close()
-	count := 0
-	for rows.Next() {
-		values := make([]any, len(cols))
-		ptrs := make([]any, len(cols))
-		for i := range values {
-			ptrs[i] = &values[i]
-		}
-		if err := rows.Scan(ptrs...); err != nil {
-			return TableManifest{}, fmt.Errorf("scan table %s: %w", table, err)
-		}
-		row := make(map[string]any, len(cols))
-		blobs := make(map[string]string)
-		for i, col := range cols {
-			row[col] = exportValue(values[i])
-			if blob, ok := values[i].([]byte); ok {
-				blobs[col] = string(blob)
-			}
-		}
-		if filter != nil {
-			keep, err := filter(table, row)
-			if err != nil {
-				return TableManifest{}, fmt.Errorf("filter table %s: %w", table, err)
-			}
-			if !keep {
-				continue
-			}
-		}
-		if filterTx != nil {
-			keep, err := filterTx(ctx, tx, table, row)
-			if err != nil {
-				return TableManifest{}, fmt.Errorf("filter table %s in export transaction: %w", table, err)
-			}
-			if !keep {
-				continue
-			}
-		}
-		data, err := encodeSnapshotRow(row, blobs)
-		if err != nil {
-			return TableManifest{}, fmt.Errorf("encode table %s: %w", table, err)
-		}
-		if err := writer.rotateIfNeeded(); err != nil {
-			return TableManifest{}, err
-		}
-		if _, err := writer.Write(data); err != nil {
-			return TableManifest{}, fmt.Errorf("write table %s: %w", table, err)
-		}
-		count++
-		if err := writer.finishRow(); err != nil {
-			return TableManifest{}, err
-		}
-	}
-	if err := rows.Err(); err != nil {
-		return TableManifest{}, err
-	}
-	if err := writer.close(); err != nil {
-		return TableManifest{}, err
-	}
-	return TableManifest{Name: table, Files: writer.files, FileManifests: writer.fileManifests, Columns: cols, Rows: count}, nil
-}
-
-func importTable(ctx context.Context, tx *sql.Tx, rootDir string, table TableManifest, filter RowFilter, importRow RowImportFunc, progress func(ImportProgress)) (int, error) {
-	files, err := importFileManifests(table)
-	if err != nil {
-		return 0, err
-	}
-	if len(files) == 0 {
-		return 0, nil
-	}
-	reportImportProgress(progress, ImportProgress{Phase: "table_start", Table: table.Name, FileCount: len(files), TotalRows: table.Rows})
-	totalRows := 0
-	for index, entry := range files {
-		rel := entry.Path
-		path, err := confinedSnapshotFile(rootDir, rel)
-		if err != nil {
-			return totalRows, err
-		}
-		file, err := os.Open(path)
-		if err != nil {
-			return totalRows, fmt.Errorf("open %s: %w", rel, err)
-		}
-		fileProgress := ImportProgress{Phase: "file_start", Table: table.Name, File: rel, FileIndex: index + 1, FileCount: len(files), TotalRows: table.Rows}
-		reportImportProgress(progress, fileProgress)
-		var expected *FileManifest
-		if len(table.FileManifests) > 0 {
-			expected = &entry
-		}
-		rows, err := importJSONLGzip(ctx, tx, file, table.Name, filter, importRow, expected)
-		if err != nil {
-			_ = file.Close()
-			return totalRows, fmt.Errorf("import %s: %w", rel, err)
-		}
-		if err := file.Close(); err != nil {
-			return totalRows, fmt.Errorf("close %s: %w", rel, err)
-		}
-		totalRows += rows
-		fileProgress.Phase = "file_done"
-		fileProgress.Rows = rows
-		reportImportProgress(progress, fileProgress)
-	}
-	return totalRows, nil
-}
-
-func importJSONLGzip(ctx context.Context, tx *sql.Tx, reader io.Reader, table string, filter RowFilter, importRow RowImportFunc, expected *FileManifest) (int, error) {
-	hasher := sha256.New()
-	counter := &countingWriter{w: hasher}
-	if expected != nil {
-		reader = io.TeeReader(reader, counter)
-	}
-	gz, err := gzip.NewReader(reader)
-	if err != nil {
-		return 0, fmt.Errorf("open gzip for %s: %w", table, err)
-	}
-	defer gz.Close()
-	scanner := bufio.NewScanner(gz)
-	scanner.Buffer(make([]byte, 0, 1024*1024), 64*1024*1024)
-	rows := 0
-	decodedRows := 0
-	for scanner.Scan() {
-		row, err := decodeSnapshotRow(scanner.Bytes(), false)
-		if err != nil {
-			return rows, fmt.Errorf("decode %s row: %w", table, err)
-		}
-		decodedRows++
-		if len(row) == 0 {
-			continue
-		}
-		if filter != nil {
-			keep, err := filter(table, row)
-			if err != nil {
-				return rows, fmt.Errorf("filter %s row: %w", table, err)
-			}
-			if !keep {
-				continue
-			}
-		}
-		importFunc := importRow
-		if importFunc == nil {
-			importFunc = insertRow
-		}
-		if err := importFunc(ctx, tx, table, row); err != nil {
-			return rows, err
-		}
-		rows++
-	}
-	// Scan through gzip EOF: Close alone does not verify the gzip checksum.
-	if err := scanner.Err(); err != nil {
-		return rows, fmt.Errorf("scan %s rows: %w", table, err)
-	}
-	if expected != nil {
-		if expected.Size != 0 && counter.n != expected.Size {
-			return rows, fmt.Errorf("compressed size mismatch: got %d, want %d", counter.n, expected.Size)
-		}
-		if expected.SHA256 != "" && !strings.EqualFold(hex.EncodeToString(hasher.Sum(nil)), expected.SHA256) {
-			return rows, errors.New("compressed SHA256 mismatch")
-		}
-		if decodedRows != expected.Rows {
-			return rows, fmt.Errorf("row count mismatch: got %d, want %d", decodedRows, expected.Rows)
-		}
-	}
-	return rows, nil
-}
-
-func reportImportProgress(progress func(ImportProgress), event ImportProgress) {
-	if progress != nil {
-		progress(event)
-	}
-}
-
-func deleteImportTable(ctx context.Context, tx *sql.Tx, table string, deleteTable DeleteFunc) error {
-	if deleteTable != nil {
-		return deleteTable(ctx, tx, table)
-	}
-	if _, err := tx.ExecContext(ctx, "delete from "+store.QuoteIdent(table)); err != nil {
-		return fmt.Errorf("clear table %s: %w", table, err)
-	}
-	return nil
-}
-
-func insertRow(ctx context.Context, tx *sql.Tx, table string, row map[string]any) error {
-	cols := make([]string, 0, len(row))
-	for col := range row {
-		cols = append(cols, col)
-	}
-	sort.Strings(cols)
-	quoted := make([]string, 0, len(cols))
-	holders := make([]string, 0, len(cols))
-	args := make([]any, 0, len(cols))
-	for _, col := range cols {
-		quoted = append(quoted, store.QuoteIdent(col))
-		holders = append(holders, "?")
-		args = append(args, row[col])
-	}
-	stmt := fmt.Sprintf(
-		"insert or replace into %s(%s) values(%s)",
-		store.QuoteIdent(table),
-		strings.Join(quoted, ","),
-		strings.Join(holders, ","),
-	)
-	if _, err := tx.ExecContext(ctx, stmt, args...); err != nil {
-		return fmt.Errorf("insert %s row: %w", table, err)
-	}
-	return nil
-}
-
-type shardWriter struct {
-	root          *os.Root
-	created       func(string)
-	relDir        string
-	maxShardBytes int64
-	nextShard     int
-	rowsInShard   int
-	files         []string
-	fileManifests []FileManifest
-	currentRel    string
-	file          *os.File
-	counter       *countingWriter
-	hasher        hash.Hash
-	gz            *gzip.Writer
-}
-
-func (w *shardWriter) Write(p []byte) (int, error) {
-	if w.gz == nil {
-		if err := w.open(); err != nil {
-			return 0, err
-		}
-	}
-	return w.gz.Write(p)
-}
-
-func (w *shardWriter) open() error {
-	rel := filepath.ToSlash(filepath.Join(w.relDir, fmt.Sprintf("%06d.jsonl.gz", w.nextShard)))
-	file, err := w.root.OpenFile(filepath.FromSlash(rel), os.O_WRONLY|os.O_CREATE|os.O_EXCL, 0o600)
-	if err != nil {
-		return fmt.Errorf("create %s: %w", rel, err)
-	}
-	w.created(rel)
-	w.nextShard++
-	w.rowsInShard = 0
-	w.files = append(w.files, rel)
-	w.currentRel = rel
-	w.file = file
-	w.hasher = sha256.New()
-	w.counter = &countingWriter{w: io.MultiWriter(file, w.hasher)}
-	w.gz = gzip.NewWriter(w.counter)
-	return nil
-}
-
-func (w *shardWriter) rotateIfNeeded() error {
-	if w.maxShardBytes <= 0 || w.rowsInShard == 0 || w.counter == nil || w.counter.n < w.maxShardBytes {
-		return nil
-	}
-	if err := w.close(); err != nil {
-		return err
-	}
-	return w.open()
-}
-
-func (w *shardWriter) finishRow() error {
-	w.rowsInShard++
-	if w.maxShardBytes > 1024*1024 && w.rowsInShard%1024 != 0 {
-		return nil
-	}
-	if w.gz == nil {
-		return nil
-	}
-	return w.gz.Flush()
-}
-
-func (w *shardWriter) close() error {
-	var closeErr error
-	if w.gz != nil {
-		if err := w.gz.Close(); err != nil {
-			closeErr = err
-		}
-		w.gz = nil
-	}
-	if w.file != nil {
-		if err := w.file.Sync(); err != nil && closeErr == nil {
-			closeErr = err
-		}
-		if err := w.file.Close(); err != nil && closeErr == nil {
-			closeErr = err
-		}
-		w.file = nil
-	}
-	if closeErr != nil {
-		return fmt.Errorf("close shard: %w", closeErr)
-	}
-	if w.currentRel != "" && w.counter != nil && w.hasher != nil {
-		w.fileManifests = append(w.fileManifests, FileManifest{
-			Path:   w.currentRel,
-			Rows:   w.rowsInShard,
-			Size:   w.counter.n,
-			SHA256: hex.EncodeToString(w.hasher.Sum(nil)),
-		})
-	}
-	w.currentRel = ""
-	w.counter = nil
-	w.hasher = nil
-	return nil
-}
-
-type countingWriter struct {
-	w io.Writer
-	n int64
-}
-
-func (w *countingWriter) Write(p []byte) (int, error) {
-	n, err := w.w.Write(p)
-	w.n += int64(n)
-	return n, err
-}
-
-func exportValue(value any) any {
-	switch v := value.(type) {
-	case []byte:
-		return string(v)
-	default:
-		return v
-	}
-}
-
-func planTableIncrement(previous, current TableManifest, merge bool) TableImportPlan {
-	if !sameStrings(previous.Columns, current.Columns) {
-		return TableImportPlan{Table: current, Mode: TableImportReplace, Files: tableFileManifests(current), Reason: "columns changed"}
-	}
-	previousFiles := tableFileManifests(previous)
-	currentFiles := tableFileManifests(current)
-	if len(previousFiles) == 0 && len(currentFiles) == 0 {
-		return TableImportPlan{Table: current, Mode: TableImportSkip, Reason: "unchanged"}
-	}
-	if !allFilesHaveFingerprints(previousFiles) || !allFilesHaveFingerprints(currentFiles) {
-		return TableImportPlan{Table: current, Mode: TableImportReplace, Files: currentFiles, Reason: "missing file fingerprints"}
-	}
-	if sameLogicalFileManifests(current.Name, previousFiles, currentFiles) {
-		return TableImportPlan{Table: current, Mode: TableImportSkip, Reason: "unchanged"}
-	}
-	if merge {
-		return planTableMerge(previousFiles, currentFiles, current)
-	}
-	if len(currentFiles) < len(previousFiles) {
-		return TableImportPlan{Table: current, Mode: TableImportReplace, Files: currentFiles, Reason: "files removed"}
-	}
-	for i := 0; i < len(previousFiles)-1; i++ {
-		if !sameLogicalFileManifest(current.Name, previousFiles[i], currentFiles[i]) {
-			return TableImportPlan{Table: current, Mode: TableImportReplace, Files: currentFiles, Reason: "non-tail file changed"}
-		}
-	}
-	changed := make([]FileManifest, 0, len(currentFiles)-len(previousFiles)+1)
-	if len(previousFiles) > 0 {
-		oldTail := previousFiles[len(previousFiles)-1]
-		newTail := currentFiles[len(previousFiles)-1]
-		if logicalFileKey(current.Name, oldTail.Path) != logicalFileKey(current.Name, newTail.Path) {
-			return TableImportPlan{Table: current, Mode: TableImportReplace, Files: currentFiles, Reason: "tail path changed"}
-		}
-		if !sameLogicalFileManifest(current.Name, oldTail, newTail) {
-			return TableImportPlan{Table: current, Mode: TableImportReplace, Files: currentFiles, Reason: "tail file changed"}
-		}
-	}
-	for i := len(previousFiles); i < len(currentFiles); i++ {
-		changed = append(changed, currentFiles[i])
-	}
-	if len(changed) == 0 {
-		return TableImportPlan{Table: current, Mode: TableImportSkip, Reason: "unchanged"}
-	}
-	return TableImportPlan{Table: current, Mode: TableImportFiles, Files: changed, Reason: "tail files changed"}
-}
-
-func planTableMerge(previousFiles, currentFiles []FileManifest, current TableManifest) TableImportPlan {
-	previousByPath := make(map[string]FileManifest, len(previousFiles))
-	for _, file := range previousFiles {
-		previousByPath[logicalFileKey(current.Name, file.Path)] = file
-	}
-	currentPaths := make(map[string]struct{}, len(currentFiles))
-	changed := make([]FileManifest, 0, len(currentFiles))
-	for _, file := range currentFiles {
-		key := logicalFileKey(current.Name, file.Path)
-		currentPaths[key] = struct{}{}
-		previous, ok := previousByPath[key]
-		if !ok || !sameLogicalFileManifest(current.Name, previous, file) {
-			changed = append(changed, file)
-		}
-	}
-	for _, file := range previousFiles {
-		if _, ok := currentPaths[logicalFileKey(current.Name, file.Path)]; !ok {
-			return TableImportPlan{Table: current, Mode: TableImportReplace, Files: currentFiles, Reason: "files removed"}
-		}
-	}
-	if len(changed) == 0 {
-		return TableImportPlan{Table: current, Mode: TableImportSkip, Reason: "unchanged"}
-	}
-	return TableImportPlan{Table: current, Mode: TableImportFiles, Files: changed, Reason: "merge changed files"}
-}
-
-func confinedSnapshotFile(rootDir, rel string) (string, error) {
-	local := filepath.FromSlash(rel)
-	// Keep literal filesystem names; this checks traversal, not symlink targets.
-	if !filepath.IsLocal(local) || filepath.Clean(local) == "." || strings.ContainsRune(local, 0) {
-		return "", fmt.Errorf("snapshot file path %q is not under the snapshot root", rel)
-	}
-	return filepath.Join(rootDir, local), nil
-}
-
-func tableShardDir(table string) (string, error) {
-	table = strings.TrimSpace(table)
-	if table == "" {
-		return "", fmt.Errorf("table name is required")
-	}
-	if filepath.IsAbs(table) || strings.ContainsAny(table, `/\`) || table == "." || table == ".." || strings.Contains(table, "\x00") {
-		return "", fmt.Errorf("table name %q is not safe for snapshot paths", table)
-	}
-	return filepath.ToSlash(filepath.Join("tables", table)), nil
-}
-
-func tableFileManifests(table TableManifest) []FileManifest {
-	if len(table.FileManifests) > 0 {
-		out := make([]FileManifest, len(table.FileManifests))
-		copy(out, table.FileManifests)
-		return out
-	}
-	files := table.Files
-	if len(files) == 0 && strings.TrimSpace(table.File) != "" {
-		files = []string{table.File}
-	}
-	out := make([]FileManifest, 0, len(files))
-	for _, file := range files {
-		out = append(out, FileManifest{Path: file})
-	}
-	return out
-}
-
-func importFileManifests(table TableManifest) ([]FileManifest, error) {
-	if table.Rows < 0 {
-		return nil, fmt.Errorf("table %s: negative row count", table.Name)
-	}
-	entries := tableFileManifests(table)
-	modern := len(table.FileManifests) > 0
-	remainingRows := table.Rows
-	byPath := make(map[string]FileManifest, len(entries))
-	for _, file := range entries {
-		// Use the same lexical path identity as opening the shard.
-		key, err := confinedSnapshotFile(".", file.Path)
-		if err != nil {
-			return nil, fmt.Errorf("table %s: %w", table.Name, err)
-		}
-		if _, exists := byPath[key]; exists {
-			return nil, fmt.Errorf("table %s: duplicate file manifest path %q", table.Name, file.Path)
-		}
-		byPath[key] = file
-		if modern {
-			if file.Rows < 0 {
-				return nil, fmt.Errorf("table %s: negative row count for %q", table.Name, file.Path)
-			}
-			// Subtract from the declared total to avoid overflowing a sum of rows.
-			if file.Rows > remainingRows {
-				return nil, fmt.Errorf("table %s: row count does not match file manifests", table.Name)
-			}
-			remainingRows -= file.Rows
-		}
-	}
-	if !modern {
-		// Legacy table totals remain advisory, including positive values.
-		return entries, nil
-	}
-	if remainingRows != 0 {
-		return nil, fmt.Errorf("table %s: row count does not match file manifests", table.Name)
-	}
-	paths := table.Files
-	if table.File != "" {
-		key, err := confinedSnapshotFile(".", table.File)
-		if err != nil {
-			return nil, fmt.Errorf("table %s: %w", table.Name, err)
-		}
-		if len(paths) == 0 {
-			paths = []string{table.File}
-		} else if len(paths) != 1 {
-			return nil, fmt.Errorf("table %s: inconsistent file paths", table.Name)
-		} else if listedKey, err := confinedSnapshotFile(".", paths[0]); err != nil || listedKey != key {
-			return nil, fmt.Errorf("table %s: inconsistent file paths", table.Name)
-		}
-	}
-	if len(paths) == 0 {
-		paths = fileManifestPaths(table.FileManifests)
-	}
-	files := make([]FileManifest, 0, len(paths))
-	for _, path := range paths {
-		key, err := confinedSnapshotFile(".", path)
-		if err != nil {
-			return nil, fmt.Errorf("table %s: %w", table.Name, err)
-		}
-		file, ok := byPath[key]
-		if !ok {
-			return nil, fmt.Errorf("table %s: missing file manifest for %q", table.Name, path)
-		}
-		file.Path = path
-		files = append(files, file)
-		delete(byPath, key)
-	}
-	if len(byPath) != 0 {
-		return nil, fmt.Errorf("table %s: file manifests do not match file paths", table.Name)
-	}
-	return files, nil
-}
-
-func selectImportFiles(plan TableImportPlan) (TableManifest, error) {
-	table := plan.Table
-	files, err := importFileManifests(table)
-	if err != nil {
-		return TableManifest{}, err
-	}
-	modern := len(table.FileManifests) > 0
-	byPath := make(map[string]FileManifest, len(files))
-	for _, file := range files {
-		key, _ := confinedSnapshotFile(".", file.Path) // Validated above.
-		byPath[key] = file
-	}
-	for _, file := range plan.Files {
-		key, err := confinedSnapshotFile(".", file.Path)
-		if err != nil {
-			return TableManifest{}, fmt.Errorf("table %s: %w", table.Name, err)
-		}
-		expected, ok := byPath[key]
-		expected.Path = file.Path
-		if !ok || (modern && !sameFileManifest(file, expected)) {
-			return TableManifest{}, fmt.Errorf("table %s: inconsistent planned file manifest for %q", table.Name, file.Path)
-		}
-		delete(byPath, key)
-	}
-	if modern {
-		table.FileManifests = plan.Files
-	}
-	// Legacy plans synthesize zero-valued metadata; those rows are unknown.
-	table.File = ""
-	table.Files = fileManifestPaths(plan.Files)
-	table.Rows = fileManifestRows(plan.Files)
-	return table, nil
-}
-
-func allFilesHaveFingerprints(files []FileManifest) bool {
-	for _, file := range files {
-		if file.Path == "" || file.SHA256 == "" {
-			return false
-		}
-	}
-	return true
-}
-
-func sameFileManifest(a, b FileManifest) bool {
-	return a.Path == b.Path && a.Rows == b.Rows && a.Size == b.Size && a.SHA256 == b.SHA256
-}
-
-func fileManifestPaths(files []FileManifest) []string {
-	paths := make([]string, 0, len(files))
-	for _, file := range files {
-		paths = append(paths, file.Path)
-	}
-	return paths
-}
-
-func fileManifestRows(files []FileManifest) int {
-	rows := 0
-	for _, file := range files {
-		rows += file.Rows
-	}
-	return rows
-}
-
-func sameStrings(a, b []string) bool {
-	if len(a) != len(b) {
-		return false
-	}
-	for i := range a {
-		if a[i] != b[i] {
-			return false
-		}
-	}
-	return true
 }
